pyproject.toml
[tool.poetry]
name = "doorbell"
version = "0.1.0"
description = ""
authors = ["Casey Mau"]

[tool.poetry.dependencies]
python = "^3.9"
PyChromecast = "^12.1.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
